Per file buffer ownership

+++ headers/cmem-cma.h
@@ -45,12 +45,12 @@
  * `numa_node` is overwritten with the actual NUMA node on which memory has been allocated.
  */
 struct cmem_cma_alloc_req {
-    __u64 size;        /**< Requested size of the DMA buffer (in bytes) */
+    __u32 size;        /**< Requested size of the DMA buffer (in bytes) */
     __u32 numa_node;   /**< (In) NUMA node to allocate from, or -1 for default
                             (Out) NUMA node actually used. */
     __u64 dma_addr;    /**< (Out) DMA (bus) address of the allocated buffer */
     __u64 mmap_offset; /**< (Out) Offset argument to 'mmap()' of this fd: buffer_id * PAGE_SIZE */
-    __s32 buffer_id;   /**< (Out) Internal ID of the allocated buffer */
+    __u32 buffer_id;   /**< (Out) Internal ID of the allocated buffer */
     __u32 reserved0;   /**< Reserved for future use */
 };
